[#172] Refactor enrichers
