+++ code/helpers/_module.mjs
@@ -1,4 +1,4 @@
 export * as interaction from "./interaction/_module.mjs";
+export * as enrichers from "./enrichers/_module.mjs";
 
 export { default as Prelocalization } from "./prelocalization.mjs";
-export { default as Enrichers } from "./enrichers.mjs";
