+++ code/utils/_module.mjs
@@ -2,6 +2,7 @@ export { default as applyStatus } from "./apply-status.mjs";
 export { default as createDefaultIdentifier } from "./create-default-identifier.mjs";
 export { default as formatNumber } from "./format-number.mjs";
 export { default as isValidIdentifier } from "./is-valid-identifier.mjs";
+export { default as parseEnricherConfig } from "./parse-enricher-config.mjs";
 export { default as parseInputDelta } from "./parse-input-delta.mjs";
 export { default as prelocalize } from "./prelocalize.mjs";
 export { default as staticId } from "./static-id.mjs";
