+++ code/utils/parse-enricher-config.mjs
@@ -1,0 +1,18 @@
+/**
+ * Parse parameters from a string.
+ * @param {string} match
+ * @returns {object}
+ */
+export default function parseEnricherConfig(match) {
+  const config = { _config: match, values: [] };
+  for (const part of match.match(/(?:[^\s"]+|"[^"]*")+/g) ?? []) {
+    if (!part) continue;
+    const [key, value] = part.split("=");
+    const valueLower = value?.toLowerCase();
+    if (value === undefined) config.values.push(key.replace(/(^"|"$)/g, ""));
+    else if (["true", "false"].includes(valueLower)) config[key] = valueLower === "true";
+    else if (Number.isNumeric(value)) config[key] = Number(value);
+    else config[key] = value.replace(/(^"|"$)/g, "");
+  }
+  return config;
+}
